Created tables and changed fields from Customer, Movie, OrderItem classes

+++ GV.DVDCentral.BL.Models/Movie.cs
@@ -15,7 +15,7 @@ namespace GV.DVDCentral.BL.Models
 
         public string? Description { get; set; }
 
-        public double Cost { get; set; }
+        public Decimal Cost { get; set; }
 
         public int  RatingId { get; set; }
